ported accessibility checks and messages - this fixes 5 tests
git-svn-id: https://jtidy.svn.sourceforge.net/svnroot/jtidy/branches/CodeUpdateAndJava5@1233 58265ab0-d70d-0410-8993-b38ad8f1ef24

+++ src/main/java/org/w3c/tidy/Lexer.java
@@ -202,7 +202,7 @@ public class Lexer
     /**
      * for accessibility errors.
      */
-    protected short badAccess;
+    protected int badAccess;
 
     /**
      * for bad style errors.
